Refactor app.js to use routers for assignments and accounts, removing inline route handlers for improved organization and maintainability.
